Include debug and coverage build
