+++ frontend/package.json
@@ -5,10 +5,10 @@
   "main": "index.js",
   "scripts": {
     "lint": "eslint src/**/*.ts",
-    "build": "yarn run sass; rollup -c rollup.config.js",
+    "build": "yarn run sass; rollup -c rollup.config.js; SKIP_COMPRESS_JS=true rollup -c rollup.config.js; COVERAGE=true rollup -c rollup.config.js",
     "sass": "sass --no-source-map style/file_form.scss ../django_file_form/static/file_form/file_form.css",
     "tsc": "tsc --noEmit",
-    "watch": "webpack --mode development --watch"
+    "watch": "rollup -c -w rollup.config.js"
   },
   "repository": {
     "type": "git",
@@ -39,7 +39,7 @@
     "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-react": "^7.20.0",
     "rollup": "^2.32.0",
-    "rollup-plugin-istanbul": "Philipp-M/rollup-plugin-istanbul",
+    "rollup-plugin-istanbul": "^2.0.1",
     "rollup-plugin-node-polyfills": "^0.2.1",
     "rollup-plugin-serve": "^1.0.4",
     "rollup-plugin-terser": "^7.0.2",
