+++ frontend/rollup.config.js
@@ -1,27 +1,57 @@
 import babel from "@rollup/plugin-babel";
 import commonjs from "@rollup/plugin-commonjs";
+import coverage from "rollup-plugin-istanbul";
 import nodePolyfills from "rollup-plugin-node-polyfills";
 import resolve from "@rollup/plugin-node-resolve";
 import { terser } from "rollup-plugin-terser";
 
+const skipCompressJs = Boolean(process.env.SKIP_COMPRESS_JS);
+const includeCoverage = Boolean(process.env.COVERAGE);
+
+const minimize = !skipCompressJs && !includeCoverage;
+
+const getOutputFilename = () => {
+  if (includeCoverage) {
+    return "file_form.coverage.js";
+  } else if (skipCompressJs) {
+    return "file_form.debug.js";
+  } else {
+    return "file_form.js";
+  }
+};
+
+const plugins = [
+  resolve({ browser: true, extensions: [".js", ".ts"] }),
+  commonjs(),
+  nodePolyfills(),
+  babel({
+    babelHelpers: "runtime",
+    extensions: [".js", ".ts"],
+    exclude: "node_modules/core-js/**"
+  })
+];
+
+if (minimize) {
+  const terserPlugin = terser();
+  plugins.push(terserPlugin);
+}
+
+if (includeCoverage) {
+  const coveragePlugin = coverage({
+    esModules: true
+  });
+  plugins.push(coveragePlugin);
+}
+
 export default {
   input: "src/file_form.ts",
   output: {
-    file: "../django_file_form/static/file_form/file_form.js",
+    file: `../django_file_form/static/file_form/${getOutputFilename()}`,
     format: "iife",
     name: "file_form",
+    sourcemap: true,
     strict: false
   },
   external: ["jquery"],
-  plugins: [
-    resolve({ browser: true, extensions: [".js", ".ts"] }),
-    commonjs(),
-    nodePolyfills(),
-    babel({
-      babelHelpers: "runtime",
-      extensions: [".js", ".ts"],
-      exclude: "node_modules/core-js/**"
-    }),
-    terser()
-  ]
+  plugins
 };
